+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **0.7.5**
 - Initial support for sending args and kwargs for traceback frames.
+- Optimization to send the access token in a header.
 
 **0.7.4**
 - Level kwarg added to `rollbar.report_exc_info()` ([#22](https://github.com/rollbar/pyrollbar/pull/22))